Hard Coded Info contains the simple firmware for users that want to experiment and appreciate smaller code at the expense of features.
The "Hard Coded Info" typ of firmware is intended for experimenters that need a lighter firmware and are OK with having their call sign hard coded in the source code.  It does not support the Serial API that allows the the use of a PC software to set all the data and it only transmits on a single frequency.