Fixed problem with Arduino and PC program not in sync. Added debug option
There is now a debug option that enables a user to send and receive raw API data . Useful when debugging new options or to manually set stuff. E.g to force the transmitter to temporary use Low pass filter B - send the command "[CSL] S B".
I also fixed a long standing problem where the Arduino had a different view on what bands ere enabled compared to the PC.
This led teh user to belive a certain band was enabled for TX when it was not. The PC Software now queries the Arduino for GUI updates after each band option is ticked. This slows down the GUI when clicking to enable or disable a certain band in the WSPR beacon but I think it is not to bad.