V1.0  Added Type 3 transmissions . Auto-detect I2CAddress
Type 3 transmissions gives increased position precision with six letter Maidenhead position reports, also gives better altitude precision in case altitude is encoded in the pwr field. In that case first transmission converts altitude to power by diving the altitude in meter by 300 and round down to closest defined power. The second transmission takes the remainder and converts it to pwr by dividing by 20.
At the receive side the conversion is by taking the pwr in dBm and multiplying with 300 and adding the next message pwr in dBm * 20